(review) moved code from nwebconnectivity to nwcth
