--- internal/engine/experiment/nwebconnectivity/control.go
@@ -1,8 +1,0 @@
-package nwebconnectivity
-
-// ControlRequest is the request that we send to the control
-type ControlRequest struct {
-	HTTPRequest        string              `json:"http_request"`
-	HTTPRequestHeaders map[string][]string `json:"http_request_headers"`
-	TCPConnect         []string            `json:"tcp_connect"`
-}
